+++ package.json
@@ -15,8 +15,8 @@
     "build:admin": "node script/build.js",
     "build:check": "node -e \"const fs=require('fs'); const devFiles=fs.readdirSync('js').filter(f=>f.endsWith('.dev.js')); devFiles.forEach(f=>{const min=f.replace('.dev.js','.js'); if(!fs.existsSync('js/'+min)){console.error('Missing: '+min);process.exit(1);}; console.log('OK: '+f+' -> '+min);})\"",
     "start": "wp-scripts start --webpack-src-dir=src/blocks --output-path=build/blocks",
-    "test:e2e": "NODE_OPTIONS=--no-experimental-strip-types npx playwright test --config=playwright.config.js",
-    "test:e2e:ui": "NODE_OPTIONS=--no-experimental-strip-types npx playwright test --config=playwright.config.js --ui",
+    "test:e2e": "NODE_OPTIONS=--no-experimental-strip-types npx playwright test --config=playwright.config.mjs",
+    "test:e2e:ui": "NODE_OPTIONS=--no-experimental-strip-types npx playwright test --config=playwright.config.mjs --ui",
     "wp-env": "wp-env"
   },
   "repository": {
