+++ tests/e2e/specs/admin/document-upload.spec.js
@@ -7,7 +7,7 @@
  *
  * @see js/wp-document-revisions.dev.js
  */
-const { test, expect } = require( '@wordpress/e2e-test-utils-playwright' );
+import { test, expect } from '@wordpress/e2e-test-utils-playwright';
 
 test.describe( 'Document Upload Callback', () => {
 	let documentId;
